backend/pyproject.toml
[project]
name = "chippy-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Config + validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Supabase (DB + Storage + Auth admin)
    "supabase>=2.5.0",

    # Document processing
    "pymupdf>=1.24.0",

    # AI — LlamaIndex core
    "llama-index>=0.10.0",
    "llama-index-llms-google-genai>=0.3.0",
    "llama-index-embeddings-google-genai>=0.3.0",
    "llama-index-vector-stores-qdrant>=0.3.0",
    "llama-index-vector-stores-chroma>=0.2.0",

    # Vector DBs
    "qdrant-client>=1.9.0",
    "chromadb>=0.5.0",

    # Gemini SDK (for File API — PDF upload)
    "google-generativeai>=0.8.0",

    # HTTP client (used by Supabase client + internal calls)
    "httpx>=0.27.0",

    # Email validation for Pydantic EmailStr
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["app"]
